requirements.txt
aniso8601==8.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dateparser==0.7.6
Flask==1.1.2
Flask-JWT-Extended==3.24.1
flask-marshmallow==0.14.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
humanize==2.6.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
marshmallow==3.8.0
maya==0.6.1
pendulum==2.1.2
PyJWT==1.7.1
PyMySQL==0.10.1
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
pytzdata==2020.1
regex==2020.9.27
requests==2.24.0
six==1.15.0
snaptime==0.2.4
SQLAlchemy==1.3.19
tzlocal==2.1
urllib3==1.25.10
Werkzeug==1.0.1
